+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.0 (2017-02-19) ##
+
+* migrated all of the changes to a separated module
+* component and comands working
+
 ## 0.1 (2017-02-19) ##
 
 * init project